% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlTreatment.R
\name{controlTreatment}
\alias{controlTreatment}
\title{Control-Treatment labels}
\usage{
controlTreatment(len = c("short", "long"))
}
\arguments{
\item{len}{lenght of labels}
}
\description{
Control-Treatment labels
}
