% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativePositive.R
\name{negativePositive}
\alias{negativePositive}
\title{Negative-Positive labels}
\usage{
negativePositive(len = c("short", "long"))
}
\arguments{
\item{len}{lenght of labels}
}
\description{
Negative-Positive labels
}
