% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noYes.R
\name{noYes}
\alias{noYes}
\title{No-Yes labels}
\usage{
noYes(len = c("long", "short"), lang = c("english", "italian"))
}
\arguments{
\item{len}{lenght of labels}

\item{lang}{language}
}
\description{
No-Yes labels
}

