% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist.R
\name{recist}
\alias{recist}
\title{Recist labels}
\source{
Eisenhauer et al. (2008), New Response evaluation criteria in
solid tumours: Revised RECIST guideline (version 1.1), European Journal
of cancer 45 (2009) 228-247
}
\usage{
recist(len = c("short", "long"))
}
\arguments{
\item{len}{lenght of labels}
}
\description{
Recist (v1.1) labels for response evaluation criteria in solid tumors
}
