% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman_numbers.R
\docType{data}
\name{roman_numbers}
\alias{roman_numbers}
\title{Roman numbers from 1 to 100}
\format{
An object of class \code{character} of length 100.
}
\usage{
roman_numbers
}
\description{
Roman numbers from 1 to 100
}
\keyword{datasets}
